% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ases.R
\docType{package}
\name{ases}
\alias{ases}
\alias{ases-package}
\title{ases: Easily Load ASES-Family Packages}
\description{
An analytic toolkit comprising of a family of toolkit packages for ASES.
}
\details{
This package is used to load the following packages at a time \cr 
\describe{
  \item{\pkg{\link{asesgeo}}}{Geographic-related functionalities}
  \item{\pkg{\link{asesmso}}}{MSOffice-related functionalities}
  \item{\pkg{\link{asesrpt}}}{Reporting-related functionalities}
  \item{\pkg{\link{asestio}}}{Data I/O and transformation}
  \item{\pkg{\link{asesvis}}}{Visualization}
  \item{\pkg{\link{aseskit}}}{Generic toolbox}
}
}
\seealso{
\pkg{\link{aseskit}}
}
\author{
\strong{Creator, Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
