% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conflicts.R
\name{ases_conflicts}
\alias{ases_conflicts}
\title{Conflicts between the ases and other packages}
\usage{
ases_conflicts()
}
\description{
This function lists all the conflicts between packages in the ases
and other packages that you have loaded.
}
\details{
There are four conflicts that are deliberately ignored: \code{intersect},
\code{union}, \code{setequal}, and \code{setdiff} from \pkg{dplyr}. These functions
make the base equivalents generic, so shouldn't negatively affect any
existing code.
}
\examples{
ases_conflicts()
}
