% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ases_packages}
\alias{ases_packages}
\title{List all packages in the ases family}
\usage{
ases_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{logical, include ases in the list?}
}
\description{
List all packages in the ases family
}
\examples{
ases_packages()
}
