% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{ases_update}
\alias{ases_update}
\title{Update ases packages}
\usage{
ases_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
ases packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOptions("repos")}.}
}
\description{
This will check to see if all ases packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
ases_update()
}
}
\seealso{
\code{\link[aseskit]{dep_pkgs}}
}
